Se agregaron cosas
-Escudo habilidad (defensa)
-Camera shake (colocar)
-Carteles bad y perfect (colocar)
-Revisar economia (agrega un drop de moneda NO item, solo que sume al player visualmente)LISTO FALTA PROBAR EN CELU

+++ Assets/03. Scripts/ScriptsNuevos/Managers/ScoreUIManager.cs
@@ -33,12 +33,12 @@ public class ScoreUIManager : MonoBehaviour
         coinsText.text = "COINS: " + value;
     }
 
-    public void UpdateDamage(int value)
+    public void UpdateDamage(float value)
     {
         damageText.text = value.ToString();
     }
 
-    public void UpdateShield(int value)
+    public void UpdateShield(float value)
     {
         shieldText.text = value.ToString();
     }
